TRAP-160 Confirm that there is an email confirmation sent when a new subscription is subscribed

+++ Models/Discount.php
@@ -30,7 +30,7 @@ class Discount extends Model
         'name'
     ];
 
-    public function getDiscount(float $total): float
+    public function getDiscount(float $total): int
     {
         $type = $this->getAttributeValue('type');
         $value = $this->getAttributeValue('value');
